move giftcard functions in separate module
